fix: split caching utils
