+++ src/next/index.ts
@@ -2,3 +2,5 @@
 
 export * from "./appDir";
 export * from "./pagesDir";
+export * from "./withAppEmotionCache";
+export * from "./augmentDocumentWithEmotionCache";
