+++ src/test/apps/ssr/next.config.js
@@ -5,7 +5,27 @@ const withBundleAnalyzer = require("@next/bundle-analyzer")({
   enabled: process.env.ANALYZE === "true",
 });
 
-module.exports = withBundleAnalyzer({
+const transpileModules = require('next-transpile-modules')([
+	'@mui/material',
+	'@mui/styled-engine',
+	'@mui/system',
+	'@mui/utils',
+	'@mui/base',
+	'@mui/private-theming',
+
+  "@emotion/react",
+  "@emotion/server",
+  "@emotion/styled",
+  "@mui/icons-material",
+  "@mui/material",
+
+]);
+
+const paths = require('./tsconfig.json').compilerOptions.paths;
+const isDevServer = process.env.NODE_ENV === 'development';
+
+module.exports = 
+	transpileModules( withBundleAnalyzer({
   reactStrictMode: true,
   webpack: function (config, { isServer, webpack }) {
     config.resolve.fallback = {
@@ -20,4 +40,4 @@ module.exports = withBundleAnalyzer({
 
     return config;
   },
-});
+}));
